🔄 synced local 'skyvern-frontend/src/' with remote 'skyvern-frontend/src/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Add new UI components and functionality for task creation, including `SwitchBar`, `PromptBox`, and `TaskTemplateCard`, and update `TaskTemplates` to toggle between template types.
>
>   - **Components**:
>     - Add `SwitchBar` component in `SwitchBar.tsx` for toggling between options.
>     - Add `PromptBox` component in `PromptBox.tsx` for task generation from prompts.
>     - Add `TaskTemplateCard` component in `TaskTemplateCard.tsx` for displaying task templates.
>     - Update `SavedTaskCard` in `SavedTaskCard.tsx` to include hover effects and deletion functionality.
>   - **Functionality**:
>     - Implement task creation from prompts in `PromptBox.tsx` using mutations for task generation, saving, and running.
>     - Add hover effects to `SavedTaskCard` and `TaskTemplateCard` for better UI feedback.
>     - Add switch functionality in `TaskTemplates.tsx` to toggle between "Skyvern Templates" and "My Templates" using `SwitchBar`.
>   - **Misc**:
>     - Remove unused `DialogTrigger` import in `SavedTaskCard.tsx`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for 5e12e74539b481daf3d6ff2e122b5dda72ae6343. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
